back-server/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.0
oauthlib==3.2.2
outcome==1.2.0
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.2
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.6.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
wsproto==1.2.0
